+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [3.18.0] - 2022-01-15
+- Bug fixes related to daemon processes and yml files.
+- Refactored and improved locking logic to handle edge cases.
+- Fixed a bug with progress indicator, it should work correctly now.
+- Fixed a bug with activity alerts that was blocking alerts if user clicked on remind later.
 
 ## [3.18.0] - 2022-01-11
 - Fixes and improvements in the plugin.